PerfModel: merge category helpers into torch mapping
Keep torch op categorization in one module now that legacy compatibility maps are gone, and preserve RMSNorm as its own legacy report sheet.
