+++ TraceLens/PerfModel/extensions/pseudo_ops_perf_utils.py
@@ -103,4 +103,3 @@ def get_pseudo_op_mappings():
     }
 
     return pseudo_op_mappings
-
