+++ TraceLens/PerfModel/extensions/rmsnorm_perf_model_extensions.py
@@ -16,6 +16,7 @@ class RMSNorm(CoreRMSNorm):
 
     category = "RMSNorm"
     bwd_category = None
+    sheet_category = "RMSNorm"
 
 
 class aiter_rms_norm(RMSNorm):
