+++ TraceLens/Reporting/generate_perf_report_pytorch.py
@@ -17,7 +17,7 @@ import numpy as np
 import pandas as pd
 
 from TraceLens import NcclAnalyser, TraceToTree, TraceDiff, TreePerfAnalyzer
-from TraceLens.PerfModel.op_categories import build_sheet_category_to_op_names
+from TraceLens.PerfModel.torch_op_mapping import build_sheet_category_to_op_names
 from TraceLens.Reporting.reporting_utils import request_install
 
 
@@ -120,11 +120,11 @@ def apply_extension(perf_analyzer, extension_path):
     extension_path = os.path.abspath(extension_path)
     extension_name = os.path.splitext(os.path.basename(extension_path))[0]
 
-    from TraceLens.PerfModel.op_categories import (
+    from TraceLens.PerfModel.torch_op_mapping import (
+        OP_CATEGORY_REGISTRY,
         register_op_categories,
         register_perf_model_categories,
     )
-    from TraceLens.PerfModel.torch_op_mapping import OP_CATEGORY_REGISTRY
 
     spec = importlib.util.spec_from_file_location(extension_name, extension_path)
     extension = importlib.util.module_from_spec(spec)
